[tui] Minor refactors to stop using unwrap() with clipboard
- Clean up run.nu so that it runs correctly when a log.file doesn't
  already exist.

+++ run.nu
@@ -288,9 +288,11 @@ def run-with-crash-reporting [] {
 def log [] {
     clear
     cd tui
-    tail -f -s 5 log.txt
-    rm log.txt
+    if ('log.txt' | path exists) {
+        rm log.txt
+    }
     touch log.txt
+    tail -f -s 5 log.txt
     cd ..
 }
